requirements.txt
asgiref==3.5.2
attrs==22.1.0
Django==4.1
django-filter==22.1
djangorestframework==3.13.1
drf-spectacular==0.23.1
inflection==0.5.1
jsonschema==4.14.0
Markdown==3.4.1
pyrsistent==0.18.1
pytz==2022.2.1
PyYAML==6.0
sqlparse==0.4.2
uritemplate==4.1.1
